refactor(frontend): move App.tsx to pages/HomePage.tsx
# Conflicts:
#	packages/frontend/src/pages/HomePage.pagination.test.tsx
