+++ packages/frontend/src/main.tsx
@@ -2,11 +2,11 @@ import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import { BrowserRouter, Route, Routes, useParams } from "react-router-dom";
 import "./index.css";
-import App from "./App.tsx";
+import HomePage from "./pages/HomePage.tsx";
 import AppDetail from "./components/AppDetailPage/AppDetail.tsx";
 
 import { SessionProvider } from "@components/keycloakSession/SessionProvider.tsx";
-import { TodoPage } from "@TodoPage.tsx";
+import { TodoPage } from "@pages/TodoPage.tsx";
 
 const AppDetailWrapper = () => {
   const { slug } = useParams<{ slug: string }>();
@@ -21,7 +21,7 @@ createRoot(document.getElementById("root")!).render(
     <BrowserRouter>
       <SessionProvider>
         <Routes>
-          <Route path="/" element={<App />} />
+          <Route path="/" element={<HomePage />} />
           <Route path="/page/app/:slug" element={<AppDetailWrapper />} />
           <Route path="/todo" element={<TodoPage />} />
         </Routes>
