+++ packages/frontend/src/pages/HomePage.tsx
@@ -1,14 +1,14 @@
-import Header from "./components/Header";
-import Hero from "./components/Hero";
-import Filters from "./components/AppsGrid/Filters.tsx";
-import AppsGrid from "./components/AppsGrid/AppsGrid.tsx";
-import Pagination from "./components/AppsGrid/Pagination.tsx";
-import Footer from "./components/Footer";
-import Spinner from "./components/Spinner";
-import "./App.css";
-import type { AppCardProps } from "./components/types.ts";
+import Header from "../components/Header.tsx";
+import Hero from "../components/Hero.tsx";
+import Filters from "../components/AppsGrid/Filters.tsx";
+import AppsGrid from "../components/AppsGrid/AppsGrid.tsx";
+import Pagination from "../components/AppsGrid/Pagination.tsx";
+import Footer from "../components/Footer.tsx";
+import Spinner from "../components/Spinner.tsx";
+import "./HomePage.css";
+import type { AppCardProps } from "../components/types.ts";
 import { memo, useEffect, useMemo, useState } from "react";
-import { tsRestClient as defaultTsRestClient } from "./api/tsRestClient";
+import { tsRestClient as defaultTsRestClient } from "../api/tsRestClient.ts";
 import { getProjectsQuerySchema } from "@shared/contracts/publicRestContracts.ts";
 import { z } from "zod";
 
@@ -16,7 +16,7 @@ interface AppProps {
   tsRestClient?: typeof defaultTsRestClient;
 }
 
-const App = memo(({ tsRestClient = defaultTsRestClient }: AppProps) => {
+const HomePage = memo(({ tsRestClient = defaultTsRestClient }: AppProps) => {
   const [apps, setApps] = useState<AppCardProps[]>([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState<string | null>(null);
@@ -129,4 +129,4 @@ const App = memo(({ tsRestClient = defaultTsRestClient }: AppProps) => {
   );
 });
 
-export default App;
+export default HomePage;
